--- code/algorithm-demo/src/com/noob/algorithm/listnode/ListNode.java
@@ -1,20 +1,0 @@
-package com.noob.algorithm.listnode;
-
-/**
- * 链表节点定义
- */
-class ListNode {
-    int val;
-    ListNode next;
-
-    ListNode() {}
-
-    ListNode(int val) {
-        this.val = val;
-    }
-
-    ListNode(int val, ListNode next) {
-        this.val = val;
-        this.next = next;
-    }
-}
