+++ code/algorithm-demo/src/com/noob/algorithm/skill/sort/SelectionSort.java
@@ -1,4 +1,4 @@
-package com.noob.algorithm.sort;
+package com.noob.algorithm.skill.sort;
 
 public class SelectionSort {
