+++ code/algorithm-demo/src/com/noob/algorithm/skill/tree/TreeBFS.java
@@ -1,4 +1,4 @@
-package com.noob.algorithm.base.tree;
+package com.noob.algorithm.skill.tree;
 
 import java.util.ArrayList;
 import java.util.LinkedList;
