+++ code/algorithm-demo/src/com/noob/algorithm/skill/tree/TreeBaseDemo.java
@@ -1,4 +1,4 @@
-package com.noob.algorithm.base.tree;
+package com.noob.algorithm.skill.tree;
 
 
 /**
